requirements.txt
scikit-learn==0.24.2
tornado==6.1
bottle==0.12.19
python-dotenv==0.15.0
pandas==1.2.4
SQLAlchemy==1.3.23